refactored code, moved DeleteJob extension into JobExecutionExtension class and changed behaviour for WorklfowGraphNotFoundException
